refactor(editor): 拆分 EditorPane 測試檔並抽共用 helper
原本有兩個同名 EditorPane.test.tsx（root 502 行 / __tests__ 2033 行），
且兩者的 mock 世界不同、無法合併：

- __tests__ 版跑真 i18n catalogue + 真 RenamePopover，斷言鎖的是英文字串
  （'Saved toast.txt'、'No FS backend is available for this file.'）
- root 版把 presentation 全 stub 掉（i18n → key identity、RenamePopover →
  props probe、portal inline），斷言鎖的是 i18n key（'editor.save.saved'）

合併任一邊都要改斷言，所以維持兩個世界、各自依主題拆檔，全部搬進 __tests__/。

真 catalogue 世界（60 tests）：
- EditorPane.load.test.tsx (16)
- EditorPane.save.test.tsx (21)
- EditorPane.rename-recent.test.tsx (10)
- EditorPane.view-mode.test.tsx (10)
- EditorPane.host-binding.test.tsx (3)

stub 世界（14 tests）：
- EditorPane.buffer-switch.test.tsx (5)
- EditorPane.save-outcome.test.tsx (7)
- EditorPane.error-binding.test.tsx (2)

共用 helper 各一組（mocks / harness 分兩檔）：
- editor-pane-mocks.tsx + editor-pane-harness.tsx
- editor-pane-stub-mocks.tsx + editor-pane-stub-harness.tsx

mocks 與 harness 必須分檔：vi.mock factory 執行時 ../EditorPane 還在 evaluate
中，factory 若 await import 一個會 import EditorPane 的模組就會 deadlock
（實測整個 vitest 卡死）。因此 mocks 檔一律不 import 受測元件。

去重內容：7 份重複 beforeEach → resetEditorPaneStores()、兩份 pressCmdS →
一份（帶 surface 參數）、renderLoaded / renderUntitled / renderEditorPane、
backend fixture（含 statMissingUntilWritten）、seedLoadedBuffer /
seedUntitledBuffer / seedTab。

原本單一的 describe('EditorPane') 30 個 it 依主題分到 4 個檔，每檔沿用同名
describe，因此完整測試名稱完全不變。

驗證：
- 247 行 expect 斷言與舊檔逐字 diff 完全相同
- 所有 describe / it 標題完全相同（只多出 3 個同名 describe('EditorPane') 標頭）
- 測試總數 4060 未變（檔案數 346 → 352）
- lint / build 綠

eslint.config.js：新增一條僅限 __tests__/**/*.tsx 的 override 關掉
react-refresh/only-export-components — 這些 fixture 檔本來就要同時 export
stub component 與 spy / helper，而 __tests__ 下的東西不會被 hot reload。
